requirements.txt
# ============================================
# 股票分析助手 - 依赖包列表
# 安装命令：pip install -r requirements.txt
# ============================================

# --------------------------------------------
# 核心框架
# --------------------------------------------
langgraph>=0.0.50
langchain-core>=0.1.0
langchain-openai>=0.0.5

# --------------------------------------------
# 数据源（Tushare Pro - 股票数据）
# 注册地址：https://tushare.pro/register
# --------------------------------------------
tushare>=1.2.89
pandas>=2.0.0

# --------------------------------------------
# 向量检索（知识库）
# --------------------------------------------
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# --------------------------------------------
# 可选依赖
# --------------------------------------------

# 网络搜索（可选）
# tavily-python>=0.3.0

# 环境变量
python-dotenv>=1.0.0

# 数据处理
numpy>=1.24.0

# --------------------------------------------
# 开发依赖（可选）
# --------------------------------------------
# pytest>=7.0.0
# black>=23.0.0
